got the checkers to work
